refactored views.py into chunks
