+++ pharmcare/admin.py
@@ -24,6 +24,7 @@ class PatientDetailModelAdmin(admin.ModelAdmin):
     """ A class model admin that modifies/overide the admin panel of patient detail """
     model = PatientDetail
     fields = (
+        'organization',
         'first_name',
         'last_name',
         'email',
@@ -61,7 +62,8 @@ class PatientDetailModelAdmin(admin.ModelAdmin):
         'social_history',
         'slug',
         'phone_number',
-        'consultation'
+        'consultation',
+        'organization',
     ]
 
     list_per_page = 10
