+++ pharmcare/templates/pharmcare/patient-info-list.html
@@ -23,7 +23,7 @@
         <h1 class="text-3xl text-gray-800">Patient's Information</h1>
       </div>
 
-        {% if request.user.is_organizer or request.user.is_agent %}
+        {% if request.user.is_organizer or request.user.is_pharmacist %}
         <div class="text-gray-500 hover:text-blue-500">
           <a href="{% url 'pharmcare:patients-info-create' %}">Go to Patient's Create Page</a> 
         </div>
